Move augeasnode_test tests to more generic parsernode_test
